+++ rbx_types/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased Changes
 
+* Added `Content::into_value` to support converting a `Content` into its underlying value. ([#507])
+
+[#507]: https://github.com/rojo-rbx/rbx-dom/pull/507
+
 ## 2.0.0 (2025-03-28)
 * Changed `Content` to more closely align with Roblox's new `Content` type. This is a breaking change. ([#495])
 * Renamed the old `Content` to `ContentId` to reflect Roblox's API changes. ([#495])